extract midi clock tracker module
