+++ src/lib.rs
@@ -1,9 +1,11 @@
 mod beat_number;
 mod line_launcher;
 mod message;
+mod midi_clock_tracker;
 mod result;
 
 pub use beat_number::BeatNumber;
 pub use line_launcher::LineLauncher;
 pub use message::Message;
+pub use midi_clock_tracker::MidiClockTracker;
 pub use result::Result;
